+++ app/src/androidTest/java/com/android/voyageur/ui/trip/activities/ActivitiesForOneDayScreenTest.java
@@ -1,0 +1,4 @@
+package com.android.voyageur.ui.trip.activities;
+
+public class ActivitiesForOneDayScreenTest {
+}
