+++ app/src/androidTest/java/com/android/voyageur/ui/trip/schedule/ByDayScreenTest.kt
@@ -5,6 +5,7 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.onNodeWithText
 import com.android.voyageur.model.trip.Trip
+import com.android.voyageur.model.trip.TripsViewModel
 import com.android.voyageur.ui.navigation.LIST_TOP_LEVEL_DESTINATION
 import com.android.voyageur.ui.navigation.NavigationActions
 import org.junit.Before
@@ -16,36 +17,28 @@ class ByDayScreenTest {
   private val sampleTrip = Trip(name = "Sample Trip")
 
   private lateinit var navigationActions: NavigationActions
+  private lateinit var tripsViewModel: TripsViewModel
 
   @get:Rule val composeTestRule = createComposeRule()
 
   @Before
   fun setUp() {
     navigationActions = mock(NavigationActions::class.java)
+    tripsViewModel = mock(TripsViewModel::class.java)
   }
 
   @Test
   fun hasRequiredComponents() {
-    composeTestRule.setContent { ByDayScreen(sampleTrip, navigationActions) }
+    composeTestRule.setContent { ByDayScreen(tripsViewModel, sampleTrip, navigationActions) }
     composeTestRule.onNodeWithTag("byDayScreen").assertIsDisplayed()
     composeTestRule.onNodeWithTag("bottomNavigationMenu").assertIsDisplayed()
     // composeTestRule.onNodeWithTag("emptyByDayPrompt").assertIsDisplayed()
   }
-  /*
-   @Test
-   fun displaysCorrectTripName() {
-     composeTestRule.setContent { ByDayScreen(sampleTrip, navigationActions) }
-     composeTestRule
-         .onNodeWithTag("emptyByDayPrompt")
-         .assertTextContains(
-             "You're viewing the ByDay screen for Sample Trip, but it's not implemented yet.")
-   }
-  */
 
   @Test
   fun displaysBottomNavigationCorrectly() {
     //    tripsViewModel.selectTrip(Trip(name = "Sample Trip"))
-    composeTestRule.setContent { ByDayScreen(sampleTrip, navigationActions) }
+    composeTestRule.setContent { ByDayScreen(tripsViewModel, sampleTrip, navigationActions) }
 
     // Check that the bottom navigation menu is displayed
     composeTestRule.onNodeWithTag("bottomNavigationMenu").assertIsDisplayed()
