+++ app/src/main/java/com/android/voyageur/ui/trip/activities/Activities.kt
@@ -120,111 +120,6 @@ fun ActivitiesScreen(
       })
 }
 
-/** Composable that displays an activity item in a card view. */
-@Composable
-fun ActivityItem(
-    activity: Activity,
-) {
-  var isExpanded by remember { mutableStateOf(false) }
-
-  val dateFormat = DateTimeFormatter.ofPattern("dd/MM/yyyy", Locale.getDefault())
-  val timeFormat = DateTimeFormatter.ofPattern("hh:mm a", Locale.getDefault())
-
-  // Convert Timestamp to LocalDateTime for formatting
-  val startLocalDateTime =
-      activity.startTime.toDate().toInstant().atZone(ZoneId.systemDefault()).toLocalDateTime()
-  val endLocalDateTime =
-      activity.endTime.toDate().toInstant().atZone(ZoneId.systemDefault()).toLocalDateTime()
-
-  // Format the date and time
-  val dateFormatted = dateFormat.format(startLocalDateTime)
-  val startTimeFormatted = timeFormat.format(startLocalDateTime)
-  val endTimeFormatted = timeFormat.format(endLocalDateTime)
-
-  Card(
-      shape = MaterialTheme.shapes.medium,
-      modifier = Modifier.padding(start = 10.dp, end = 10.dp).testTag("cardItem_${activity.title}"),
-      content = {
-        Column(modifier = Modifier.padding(16.dp).fillMaxWidth()) {
-          Row(
-              horizontalArrangement = Arrangement.SpaceBetween,
-              verticalAlignment = Alignment.CenterVertically,
-              modifier = Modifier.fillMaxWidth()) {
-                Column(modifier = Modifier.weight(1f)) {
-                  Text(
-                      text = activity.title,
-                      fontWeight = FontWeight.Medium,
-                      modifier = Modifier.padding(end = 8.dp),
-                      fontSize = 16.sp)
-                  if (activity.hasStartTime() && activity.hasEndDate()) {
-                    Text(
-                        text = "$dateFormatted $startTimeFormatted - $endTimeFormatted",
-                        fontSize = 12.sp,
-                        color = Color.Gray)
-                  }
-                }
-
-                Row(verticalAlignment = Alignment.CenterVertically) {
-                  // Delete icon, visible only when expanded
-                  AnimatedVisibility(visible = isExpanded) {
-                    IconButton(onClick = { /*TODO: delete activity*/}) {
-                      Icon(
-                          imageVector = Icons.TwoTone.Delete,
-                          contentDescription = "Delete Activity",
-                          tint = Color.Red)
-                    }
-                  }
-
-                  // Expand/collapse icon
-                  IconButton(
-                      onClick = { isExpanded = !isExpanded },
-                      modifier = Modifier.testTag("expandIcon_${activity.title}")) {
-                        Icon(
-                            imageVector = Icons.Default.ArrowDropDown,
-                            contentDescription = if (isExpanded) "Collapse" else "Expand",
-                            modifier = Modifier.rotate(if (isExpanded) 180f else 0f))
-                      }
-                }
-              }
-
-          // Expandable content with animation
-          AnimatedVisibility(visible = isExpanded, enter = fadeIn(), exit = fadeOut()) {
-            Column(modifier = Modifier.fillMaxWidth().padding(top = 8.dp)) {
-              if (activity.hasDescription()) {
-                Text(
-                    text = "Description",
-                    fontWeight = FontWeight.SemiBold,
-                    fontSize = 14.sp,
-                    modifier = Modifier.padding(vertical = 2.dp))
-                Text(
-                    text = activity.description,
-                    fontSize = 14.sp,
-                    modifier = Modifier.padding(bottom = 8.dp))
-              }
-              // TODO: Add location once we have the final location model
-              Text(
-                  text = "Price",
-                  fontWeight = FontWeight.SemiBold,
-                  fontSize = 14.sp,
-                  modifier = Modifier.padding(vertical = 2.dp))
-              Text(
-                  text = "${activity.estimatedPrice} CHF",
-                  fontSize = 14.sp,
-                  modifier = Modifier.padding(bottom = 8.dp))
-
-              Text(
-                  text = "Type",
-                  fontWeight = FontWeight.SemiBold,
-                  fontSize = 14.sp,
-                  modifier = Modifier.padding(vertical = 2.dp))
-              Text(text = activity.activityType.toString(), fontSize = 14.sp)
-            }
-          }
-        }
-      })
-}
-
-
 @Composable
 fun AddActivityButton(navigationActions: NavigationActions) {
     FloatingActionButton(
